+++ crates/cargo-contract/src/cmd/test_cmd.rs
@@ -31,6 +31,7 @@ pub struct TestCommand {
     /// Path to the `Cargo.toml` of the contract to test.
     #[clap(long, value_parser)]
     manifest_path: Option<PathBuf>,
+    /// Activate specific features.
     #[clap(flatten)]
     features: Features,
     /// Activate all available features.
@@ -51,9 +52,10 @@ impl TestCommand {
         // Composes `cargo test` args.
         let manifest_path = ManifestPath::try_from(self.manifest_path.as_ref())?;
         let mut args = vec![manifest_path.cargo_arg()?];
-        self.features.append_to_args(&mut args);
         if self.all_features {
             args.push("--all-features".to_string());
+        } else {
+            self.features.append_to_args(&mut args);
         }
         if !self.args.is_empty() {
             args.extend(self.args.clone());
@@ -88,12 +90,9 @@ impl TestCommand {
         let output = cmd.run()?;
         if !output.status.success() {
             anyhow::bail!(
-                "Failed to run `cargo test`{}",
-                if output.stderr.is_empty() {
-                    String::new()
-                } else {
-                    format!(": {}", String::from_utf8_lossy(&output.stderr))
-                }
+                "Failed to run `cargo test`:\n{}\n{}",
+                String::from_utf8_lossy(&output.stdout),
+                String::from_utf8_lossy(&output.stderr),
             )
         }