+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]
 
+- Accept plain string call arguments by inferring types from metadata - [2161](https://github.com/use-ink/cargo-contract/pull/2161).
+
 ## Version 6.0.0-beta
 
 ### Added